Make width & height unsigned shorts, for better fit into various APIs used.

+++ include/Game.h
@@ -28,8 +28,8 @@ private:
     Game() { };
     std::string name;
     struct {
-        unsigned int width = 1200;
-        unsigned int height = 675;
+        unsigned short width = 1200;
+        unsigned short height = 675;
         bool fullscreen = false;
         bool useDesktopSize = true;
         bool hideMouseFullscreen = true;
